Abstract class LoadComponent
