+++ src/main/java/seedu/duke/commands/AddCommand.java
@@ -2,8 +2,8 @@ package seedu.duke.commands;
 
 import seedu.duke.DukeException;
 import seedu.duke.component.Capacitor;
-import seedu.duke.component.Component;
 import seedu.duke.component.Inductor;
+import seedu.duke.component.LoadComponent;
 import seedu.duke.component.Resistor;
 import seedu.duke.ui.Ui;
 
@@ -26,7 +26,7 @@ public class AddCommand extends Command {
     @Override
     public void execute(Ui ui) throws DukeException {
         // Placeholder, to be replaced with Template.getComponent
-        Component c = getComponent(component);
+        LoadComponent c = getComponent(component);
         if (config.equals("series")) {
             System.out.println(c.addSeries(value));
         } else {
@@ -35,7 +35,7 @@ public class AddCommand extends Command {
     }
 
     // Placeholder, getComponent would come from Template class instead
-    private Component getComponent(String component) throws DukeException {
+    private LoadComponent getComponent(String component) throws DukeException {
         switch (component) {
         case "r":
             return new Resistor(1);
