+++ src/main/java/seedu/duke/component/Capacitor.java
@@ -1,6 +1,6 @@
 package seedu.duke.component;
 
-public class Capacitor extends Component {
+public class Capacitor extends LoadComponent {
     private static final String SI_UNIT = "µF";
 
     public Capacitor(double value) {
