+++ src/main/java/seedu/duke/component/Resistor.java
@@ -1,6 +1,6 @@
 package seedu.duke.component;
 
-public class Resistor extends Component {
+public class Resistor extends LoadComponent {
     private static final String SI_UNIT = "Ω";
 
     public Resistor(double value) {
