+++ src/main/java/seedu/duke/component/VoltageSource.java
@@ -1,6 +1,6 @@
 package seedu.duke.component;
 
-public class VoltageSource extends Component  {
+public class VoltageSource extends Component {
     private static final String SI_UNIT = "V";
 
     public VoltageSource(double value) {
